requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.3.27
langchain-google-genai>=2.0.10
langchain-core>=0.3.72
google-generativeai>=0.8.5
